requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.3.25
    # via -r requirements.in
langchain-core==0.3.61
    # via
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.18
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.42
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
narwhals==1.40.0
    # via altair
numpy==2.2.6
    # via
    #   pandas
    #   pydeck
    #   streamlit
openai==1.82.0
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   langchain-core
    #   langsmith
    #   pytest
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.2.1
    # via streamlit
pluggy==1.6.0
    # via pytest
protobuf==6.31.0
    # via streamlit
pyarrow==20.0.0
    # via streamlit
pydantic==2.11.5
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pytest==8.3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via langchain
streamlit==1.45.1
    # via -r requirements.in
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
tiktoken==0.9.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via openai
typing-extensions==4.13.2
    # via
    #   altair
    #   anyio
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
zstandard==0.23.0
    # via langsmith
